Remove dependency of service_client from mpe_contract and introduce payment_channel_provider to handle channel operations
